--- core/sdk/src/CandyShopModel.ts
@@ -1,4 +1,0 @@
-export enum Blockchain {
-  EMV = 'evm',
-  Solana = 'solana'
-}
