+++ core/sdk/src/index.ts
@@ -1,9 +1,8 @@
-export * from './SolanaCandyShop';
-export * from './SolanaCandyShopTrade';
 export * from './CandyShopDropAPI';
 export * from './CandyShopInfoAPI';
 export * from './CandyShopAuctionAPI';
-export * from './CandyShopModel';
-export * from './SolanaCandyShopModel';
+export * from './shop/sol/CandyShop';
+export * from './shop/sol/CandyShopTrade';
+export * from './shop/sol/CandyShopModel';
 /* Only export the vendor folder to control the exposure of some internal methods */
 export * from './vendor';
