+++ core/sdk/src/shop/sol/CandyShopModel.ts
@@ -1,5 +1,6 @@
 import { BN, Idl, Program, web3 } from '@project-serum/anchor';
 import { AnchorWallet } from '@solana/wallet-adapter-react';
+import { ShopSettings } from '../abstract/AbstractShopTypes';
 
 /**
  * Program version that CandyShop was created with
@@ -19,38 +20,9 @@ export interface CandyShopConstructorParams {
   /** web3.Cluster mainnet, devnet */
   env: web3.Cluster;
   /** Optional, additional shop settings */
-  settings?: Partial<CandyShopSettings>;
+  settings?: Partial<ShopSettings>;
   /** Indicates if this shop uses enterprise program functionality. Defaults to false */
   isEnterprise?: boolean;
-}
-
-/**
- * Custom CandyShop settings
- *
- * @property {string} currencySymbol
- * @property {number} currencyDecimals
- * @property {number} priceDecimals
- * @property {number} volumeDecimals
- * @property {string} mainnetConnectionUrl
- * @property {object | undefined} connectionConfig
- */
-export interface CandyShopSettings {
-  /** Shop transaction currency symbol (default is SOL) */
-  currencySymbol: string;
-  /** Shop transaction currency decimals (default is 9 for SOL) */
-  currencyDecimals: number;
-  /** Number of min decimals to display for price numbers (default is 0) */
-  priceDecimalsMin: number;
-  /** Number of max decimals to display for price numbers (default is 3) */
-  priceDecimals: number;
-  /** Number of min decimals to display for volume numbers (default is 0) */
-  volumeDecimalsMin: number;
-  /** Number of max decimals to display for volume numbers (default is 1) */
-  volumeDecimals: number;
-  /** Rpc connection endpoint */
-  mainnetConnectionUrl: string;
-  /** Connection config options */
-  connectionConfig: object | undefined;
 }
 
 /**
